requirements.txt
sqlalchemy>=2.0.0
pandas>=1.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandera>=0.18.0